+++ theme/js/components/search/search.vue
@@ -120,7 +120,7 @@
   </section>
   <section class="search-results fr-mt-1w fr-mt-md-3w" ref="results" v-bind="$attrs">
     <transition mode="out-in">
-      <div v-if="loading || (disableFirstSearch && !results.length)">
+      <div v-if="loading">
         <Loader />
       </div>
       <ul v-else-if="results.length">
@@ -204,14 +204,19 @@ export default {
     }
     // set all other search params as facets
     this.facets = searchParams;
-    if(!this.disableFirstSearch) {
+    if(this.disableFirstSearch) {
+      this.loading = true;
+    } else {
       this.search();
     }
   },
   mounted() {
     if(this.disableFirstSearch) {
-      this.results = JSON.parse(this.$refs.results.dataset.results);
-      this.totalResults = JSON.parse(this.$refs.results.dataset.totalResults);
+      if (this.$refs.results.dataset.totalResults > 0) {
+        this.results = JSON.parse(this.$refs.results.dataset.results);
+        this.totalResults = JSON.parse(this.$refs.results.dataset.totalResults);
+      }
+      this.loading = false;
     }
   },
   watch: {